Implement JSON parsing through gson Remove lagging in initial map
