+++ android/app/src/googleplay/AndroidManifest.xml
@@ -16,7 +16,7 @@
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="AIzaSyCdF4kS7Aw8jdVbW7PK95QfqHhyngMVkJs" />
+            android:value="@string/google_maps_api_key" />
 
     </application>
 </manifest>
