--- android/app/src/main/assets/map.json
@@ -1,143 +1,0 @@
-{
-"map":
-[
-  {
-    "floor": 3,
-    "id": 1,
-    "latitude": "52.51420",
-    "longitude": "13.38880",
-    "name": "Dalton Hall",
-    "room": null
-  },
-  {
-    "floor": 3,
-    "id": 2,
-    "latitude": "52.5220",
-    "longitude": "13.386898",
-    "name": "Lewis Lab",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 3,
-    "latitude": "52.52434",
-    "longitude": "13.389693",
-    "name": "Digital Design Studio",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 4,
-    "latitude": "52.52436",
-    "longitude": "13.387893",
-    "name": "Digital Design Studio",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 5,
-    "latitude": "52.52410",
-    "longitude": "13.388808",
-    "name": "Observatory Room",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 6,
-    "latitude": "52.52120",
-    "longitude": "13.382898",
-    "name": "Eco Garden Lab",
-    "room": null
-  },
-  {
-    "floor": 3,
-    "id": 7,
-    "latitude": "52.52439",
-    "longitude": "13.388393",
-    "name": "Planck Lab",
-    "room": null
-  },
-  {
-    "floor": 3,
-    "id": 8,
-    "latitude": "52.52420",
-    "longitude": "13.388898",
-    "name": "Faraday Lab",
-    "room": null
-  },
-  {
-    "floor": 3,
-    "id": 9,
-    "latitude": "52.52400",
-    "longitude": "13.388808",
-    "name": "Pauling Lab",
-    "room": null
-  },
-  {
-    "floor": 3,
-    "id": 10,
-    "latitude": "52.52320",
-    "longitude": "13.388198",
-    "name": "Fermi Lab",
-    "room": null
-  },
-  {
-    "floor": 2,
-    "id": 11,
-    "latitude": "52.52400",
-    "longitude": "13.388800",
-    "name": "Herschel Lab",
-    "room": null
-  },
-  {
-    "floor": 2,
-    "id": 12,
-    "latitude": "52.52300",
-    "longitude": "13.386898",
-    "name": "Einstein Room",
-    "room": null
-  },
-  {
-    "floor": 2,
-    "id": 13,
-    "latitude": "52.52415",
-    "longitude": "13.388898",
-    "name": "Einstein Hall",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 14,
-    "latitude": "52.52414",
-    "longitude": "13.388390",
-    "name": "Tinkering Studio",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 15,
-    "latitude": "52.52405",
-    "longitude": "13.388379",
-    "name": "Observatory Room",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 16,
-    "latitude": "52.52414",
-    "longitude": "13.388390",
-    "name": "Eco Garden Lab",
-    "room": null
-  },
-  {
-    "floor": 1,
-    "id": 17,
-    "latitude": "52.523594",
-    "longitude": "13.388190",
-    "name": "Ground Floor, Scientist For a day, Hall A",
-    "room": null
-  }
-]
-}
-
-
