+++ android/app/src/main/assets/microlocations.json
@@ -1,138 +1,140 @@
 [
     {
-        "floor": null, 
-        "id": 1, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 3, Level 3, Dalton Hall", 
+        "floor": 5,
+        "id": 1,
+        "latitude": "52.51420",
+        "longitude": "13.38880",
+        "name": "Dalton Hall",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 2, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 3, Lewis Lab", 
+        "floor": 3,
+        "id": 2,
+        "latitude": "52.5220",
+        "longitude": "13.386898",
+        "name": "Lewis Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 3, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Digital Design Studio", 
+        "floor": 1,
+        "id": 3,
+        "latitude": "52.52434",
+        "longitude": "13.389693",
+        "name": "Digital Design Studio",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 4, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Digital Design Studio", 
+        "floor": 1,
+        "id": 4,
+        "latitude": "52.52436",
+        "longitude": "13.387893",
+        "name": "Digital Design Studio",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 5, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Observatory Room", 
+        "floor": 1,
+        "id": 5,
+        "latitude": "52.52410",
+        "longitude": "13.388808",
+        "name": "Observatory Room",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 6, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Eco Garden Lab", 
+        "floor": 1,
+        "id": 6,
+        "latitude": "52.52120",
+        "longitude": "13.382898",
+        "name": "Eco Garden Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 7, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 3, Planck Lab", 
+        "floor": 3,
+        "id": 7,
+        "latitude": "52.52439",
+        "longitude": "13.388393",
+        "name": "Planck Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 8, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 3, Faraday Lab", 
+        "floor": 3,
+        "id": 8,
+        "latitude": "52.52420",
+        "longitude": "13.388898",
+        "name": "Faraday Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 9, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 3, Pauling Lab", 
+        "floor": 3,
+        "id": 9,
+        "latitude": "52.52400",
+        "longitude": "13.388808",
+        "name": "Pauling Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 10, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 3, Fermi Lab", 
+        "floor": 3,
+        "id": 10,
+        "latitude": "52.52320",
+        "longitude": "13.388198",
+        "name": "Fermi Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 11, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 2, Herschel Lab", 
+        "floor": 2,
+        "id": 11,
+        "latitude": "52.52400",
+        "longitude": "13.388800",
+        "name": "Herschel Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 12, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 2, Einstein Room", 
+        "floor": 2,
+        "id": 12,
+        "latitude": "52.52300",
+        "longitude": "13.386898",
+        "name": "Einstein Room",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 13, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 2, Einstein Hall", 
+        "floor": 2,
+        "id": 13,
+        "latitude": "52.52415",
+        "longitude": "13.388898",
+        "name": "Einstein Hall",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 14, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Tinkering Studio", 
+        "floor": 1,
+        "id": 14,
+        "latitude": "52.52414",
+        "longitude": "13.388390",
+        "name": "Tinkering Studio",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 15, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Observatory Room", 
+        "floor": 1,
+        "id": 15,
+        "latitude": "52.52405",
+        "longitude": "13.388379",
+        "name": "Observatory Room",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 16, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Eco Garden Lab", 
+        "floor": 1,
+        "id": 16,
+        "latitude": "52.52414",
+        "longitude": "13.388390",
+        "name": "Eco Garden Lab",
         "room": null
-    }, 
+    },
     {
-        "floor": null, 
-        "id": 17, 
-        "latitude": null, 
-        "longitude": null, 
-        "name": "Level 1, Ground Floor, Scientist For a day, Hall A", 
+        "floor": 1,
+        "id": 17,
+        "latitude": "52.523594",
+        "longitude": "13.388190",
+        "name": "Ground Floor, Scientist For a day, Hall A",
         "room": null
     }
 ]
+
+
